Development

+++ app/Http/Controllers/Pizza/PizzaController.php
@@ -9,6 +9,10 @@ use App\Http\Resources\PizzaResource;
 use App\Models\Pizza;
 use App\Models\PizzaImage;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Http\Request;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
+
 use Illuminate\Support\Str;
 
 class PizzaController extends Controller
@@ -20,14 +24,26 @@ class PizzaController extends Controller
         return PizzaResource::collection($pizza);
     }
 
-    public function store(PizzaCreateRequest $request)
+    public function store(Request $request)
     {
-        DB::beginTransaction();
+        $admin = Gate::authorize('delete', 'users');
 
+        if (!$admin) {
+            return response(['message' => 'Unauthorized'], Response::HTTP_UNAUTHORIZED);
+        }
+
+        $request->validate([
+            'name' => 'required',
+            'description' => 'required',
+            'price' => 'required',
+            'variant_id' => 'required'
+        ]);
+
+        DB::beginTransaction();
         try {
             $pizza = new Pizza();
             $pizza->name = $request->name;
-            $pizza->description = $pizza->description;
+            $pizza->description = $request->description;
             $pizza->price = $request->price;
             $pizza->variant_id = $request->variant_id;
             $pizza->save();
@@ -35,9 +51,9 @@ class PizzaController extends Controller
             if ($request->pizza_images) {
                 for ($i = 0; $i < count($request->pizza_images); $i++) {
                     $file = $request->pizza_images[$i];
-                    $imageName = time() . Str::random(10) . '.' . $request->pizza_images[$i]->extension();
-                    $path =  'images/pizzas/';
-                    $documentURL =  $path . '/' . $imageName;
+                    $imageName = time() . Str::random(4) . '.' . $request->pizza_images[$i]->extension();
+                    $path = "images/pizzas";
+                    $documentURL = $path . '/' . $imageName;
                     $file->move($path, $imageName);
 
                     PizzaImage::create([
@@ -50,18 +66,10 @@ class PizzaController extends Controller
 
             DB::commit();
             return response(['message' => 'Pizza Created Successfully']);
-            // Pizza::create([
-            //     'name' => $request->name,
-            //     'description' => $request->description,
-            //     "price" => $request->price,
-            //     'variant_id' => $request->variant_id
-            // ]);
         } catch (\Throwable $th) {
             DB::rollBack();
-            return response(['message' => 'Ooops, we are missing something']);
+            return response($th);
         }
-
-        return response(['message' => 'Pizza created successfully']);
     }
 
     public function show($id)
@@ -72,11 +80,17 @@ class PizzaController extends Controller
             return response(['message' => 'Pizza not found']);
         }
 
-        return PizzaResource::collection($pizza);
+        return new PizzaResource($pizza);
     }
 
     public function update(PizzaUpdateRequest $request, $id)
     {
+        $admin = Gate::authorize('delete', 'users');
+
+        if (!$admin) {
+            return response(['message' => 'Unauthorized'], Response::HTTP_UNAUTHORIZED);
+        }
+
         $pizza = Pizza::find($id);
 
         if (!$pizza) {
@@ -85,11 +99,19 @@ class PizzaController extends Controller
 
         $pizza->update($request->all());
 
-        return response(['message' => 'Pizza updated successfully']);
+        return response(['message' => 'Pizza updated successfully'], Response::HTTP_ACCEPTED);
     }
 
     public function destroy($id)
     {
+        $admin = Gate::authorize('delete', 'users');
+
+        if (!$admin) {
+            return response(['message' => 'Unauthorized'], Response::HTTP_UNAUTHORIZED);
+        }
+
         Pizza::destroy($id);
+
+        return response(["message" => 'Pizza deleted successfully']);
     }
 }
