+++ app/Http/Controllers/Pizza/VariantController.php
@@ -7,6 +7,8 @@ use App\Http\Requests\VariantRequest;
 use App\Http\Resources\VariantResource;
 use App\Models\Variant;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class VariantController extends Controller
 {
@@ -18,6 +20,12 @@ class VariantController extends Controller
 
     public function store(VariantRequest $request)
     {
+        $admin = Gate::authorize('delete', 'users');
+
+        if (!$admin) {
+            return response(['message' => 'Unauthorized'], Response::HTTP_UNAUTHORIZED);
+        }
+
         Variant::create($request->only('name'));
 
         return response(['message' => 'Variant Variant created']);
@@ -32,6 +40,12 @@ class VariantController extends Controller
 
     public function update(Request $request, $id)
     {
+        $admin = Gate::authorize('delete', 'users');
+
+        if (!$admin) {
+            return response(['message' => 'Unauthorized'], Response::HTTP_UNAUTHORIZED);
+        }
+
         $variant = Variant::find($id);
 
         $variant->update($request->only('name'));
@@ -41,6 +55,12 @@ class VariantController extends Controller
 
     public function destroy($id)
     {
+        $admin = Gate::authorize('delete', 'users');
+
+        if (!$admin) {
+            return response(['message' => 'Unauthorized'], Response::HTTP_UNAUTHORIZED);
+        }
+
         Variant::destroy($id);
 
         return response(['message' => 'Variant deleted successfully']);
